==> 03_splsda.Rmd <==
---
title: "03_splsda"
author: "Ashfaq Ali"
date: "01/12/2021"
output: 
  html_document:
    code_folding: hide
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, 
                      collapse = TRUE, 
                      warning = FALSE, 
                      message = FALSE, 
                      cache = FALSE, 
                      cache.path = "../cache/", 
                      fig.height = 5, 
                      fig.width = 6,
                      eval=TRUE)
```

```{r}
library(mixOmics)
```

# extract training data and name each data frame
```{r, eval=TRUE, include=FALSE}
data_all <- readRDS("../data_sumExpfilImp.rds")

rowData(data_all)$variance <- assays(data_all)$loess %>% rowVars(na.rm = TRUE)

```

# Sample level data integration vs. feature level integration.

In the sample level data integration, we investigate how individual data sets at 24h, 48h and 72h and how features(proteins) in these data sets are related to the phenotype variable. This is usually applicable when the same samples are measured using multiple technologies. This analyses can be useful effect if sample level integration required to identify correlated features from different time-points and how they are jointly correlated. The advantage here can that the same feature do not have to be measured in different studies.

Feature level integration allows integrated analyses of same features measured from different samples on the same outcome variable. The advantage of this method is that features donot have be from the same samples. 


```{r, include=FALSE,eval=FALSE}
diab_dat <- list()
X <- list(t1 = assays(data_all[, data_all$time=="-24-" & !is.na(data_all$randomisation_code)])$loess,
          t2 = assays(data_all[, data_all$time=="-48-" & !is.na(data_all$randomisation_code)])$loess, 
          t3 = assays(data_all[, data_all$time=="-72-" & !is.na(data_all$randomisation_code)])$loess)

cols <- lapply(lapply(X, colnames), function(x) str_split(x,pattern = "_", simplify = TRUE)[,1]) # remove trailing time point id

## Rename all matrices of the lins
for (i in 1:length(X)){
  colnames(X[[i]]) <- cols[[i]]
}

## common samples in 3 timepoints
cols <- Reduce(intersect, cols)

Xrand <-  lapply(X, function(x) t(x[, cols]))  ## Select common samples

y_data <- data_all[, data_all$time=="-24-" & 
                     !is.na(data_all$randomisation_code) & 
                     data_all$screening_number %in% cols
                   ] %>% colData()

Yrand <- as.factor(y_data$randomisation_code)
Ycpc <- as.factor(y_data$binarysecout180daycpcscore)
#Y3 <- as.factor(y_data$shockable)
names(Y) <- rownames(y_data)

#saveRDS(list(X,Y,Y2, y_data, cols), "../plsdat_data_diablo.rds")
```


## Randomization code

```{r}
diablo_res <- list()
diablo_res$randSplsda <- block.splsda(X = Xrand, Y = Yrand, ncomp = 5)
plotIndiv(diablo_res$randSplsda,study = 'global', ellipse = TRUE)
```

## CPC score


```{r}

diablo_res$CPCSplsda <- block.splsda(X = Xrand, Y = Ycpc, ncomp = 5)
plotIndiv(diablo_res$CPCSplsda,study = 'global', ellipse = TRUE)
```

## Shock

```{r}
y_data_s <- y_data[y_data$shockable==0 | y_data$shockable==1 & !is.na(y_data$shockable),]
Xs <- lapply(X, function(x) t(x[, y_data_s$screening_number]))
Yshock <- as.factor(y_data_s$shockable)
```


```{r}
diablo_res$ShockSplsda <- block.splsda(X = Xs, Y = Yshock, ncomp = 5)
plotIndiv(diablo_res$ShockSplsda,study = 'global', ellipse = TRUE)

```



## Tune Number of PCs and features.

### Identify prefferred number of components in each dataset

```{r}
# mpls <- list()
# perf <- paste0(opt$outcome, "perf.diablo")

set.seed(123) # for reproducibility, only when the `cpus' argument is not used
t1 <- proc.time()
perf.pls <- perf(MySplsda, validation = "Mfold", folds = 5, nrepeat = 5, cpus = 5,progressBar = T)
t2 <- proc.time()
running_time <- t2 - t1
running_time
```

### Identify optimal number of features


```{r}
design <- matrix(0.1,
  ncol = length(X), nrow = length(X),
  dimnames = list(names(X), names(X))
)
diag(design) <- 0

## Identify number of features for each dataset

list.keepX <- list(t1 = c(4:10, 12, 14, 16, 18, 20), t2 = c(4:10, 12, 14, 16, 18, 20),t3= c(4:10, 12, 14, 16, 18, 20))
# list.keepX <- list(mRNA= c(5,  seq(20, 60, 10)), protein = c(5,  seq(20, 60, 10)))


# tune <- paste0(opt$outcome, tune.diablo)
BPPARAM <- BiocParallel::MulticoreParam(workers = parallel::detectCores()-1)
BPPARAM
tune.pls <- tune.block.splsda(X, Y2,
  ncomp = 4, # we suggest to push ncomp a bit more, e.g. 4
  validation = "Mfold", dist = "max.dist", progressBar = TRUE,
  test.keepX = list.keepX, design=design,
  folds = 5, nrepeat = 3,  BPPARAM = BPPARAM
) # suggested nrepeat = 50


ncomp <- tune.pls$choice.ncomp$ncomp # optimal number of components based on t-tests on the error rate
select.keepX <- tune.pls$choice.keepX # optimal number of variables to select

# results <- paste0(opt$outcome, results.diablo)
results.dblo <- block.splsda(X = X, Y = Y2, keepX = select.keepX, ncomp = 4, max.iter = 30)

```

```{r}
plotDiablo(results.dblo)
```


```{r}
# with background
background = background.predict(results.dblo, comp.predicted=2, dist = "max.dist") 
#optional: xlim = c(-40,40), ylim = c(-30,30))

plotIndiv(results.dblo, comp = c(1,3),
           ind.names = FALSE, title = "Maximum distance",
          legend = TRUE)
```


### Report annotated used variables
```{r}
vars <- selectVar(results.dblo, comp = 1, block = c(1, 2))
vars <- list(proteins = vars$protein$value %>% rownames_to_column(), mRNA = vars$mRNA$value %>% rownames_to_column())
```



```{r}
mint.plsda.res.perf = mint.plsda(X = mint_data$protein, Y = mint_data$Y, study = mint_data$study, ncomp = 5)

set.seed(2543)  # for reproducible result in this example
## S3 method for class 'splsda'
perf(object, 
dist = c("all", "max.dist", "centroids.dist", "mahalanobis.dist"),
validation = c("Mfold", "loo"),
folds = 10, nrepeat =5, auc = FALSE, progressBar = TRUE, cpus, ...)
perf.mint.plsda.cell <- perf(mint.plsda.res.perf, validation = "Mfold", folds = 5, 
                  progressBar = FALSE, auc = TRUE) 

```


```{r, include=FALSE, eval=FALSE}
list.keepX <- list(t1 = c(10, 10, 10), t2 = c(15,10, 10), t3 = c(10, 10,10))
diablo_rand <- block.splsda(X, Y, keepX=list.keepX, ncomp = 3)
plotIndiv(diablo_rand, comp = c(1,2)) ## sample plot

plotVar(diablo_rand)
```

```{r, include=FALSE, eval=FALSE}
plotDiablo(diablo_rand)

circosPlot(diablo_rand, cutoff = 0.7)

auc_t1 <- auroc(diablo_rand, roc.block = "t1", roc.comp = 3)


auc_t2 <- auroc(diablo_rand, roc.block = "t2", roc.comp = 3)


auc_t2 <- auroc(diablo_rand, roc.block = "t3", roc.comp = 2)
```


```{r, include=FALSE, eval=FALSE}
Y2 <- as.factor(y_data$binarysecout180daycpcscore)

plsda.res.perf = mint.plsda(X = mint_data$protein, Y = mint_data$Y, study = mint_data$study, ncomp = 5)

set.seed(2543)  # for reproducible result in this example
perf.mint.plsda.cell <- perf(mint.plsda.res.perf, validation = "Mfold", folds = 5, 
                  progressBar = FALSE, auc = TRUE) 

plot(perf.mint.plsda.cell, col = color.mixo(5:7))

#names(Y) <- rownames(y_data)
summary(Y2)

list.keepX2 <- list(t1 = c(10, 10), t2 = c(15,10), t3 = c(10, 10))

diablo_cpc <- block.splsda(X, Y2, keepX=list.keepX2, ncomp = 3)

plotIndiv(diablo_cpc, ind.names = FALSE ,comp = c(1,2), ellipse = TRUE) ## sample plot
plotDiablo(diablo_cpc)
```

*********

## Multivariate approach to integrate data from three timepoints at feature level

Here we uses "mixOmics" package approach in an integrated analyses of the protein data from 3 timepoints in order to capture features from all three timepoints and their correlation with the outcome variable. 


```{r data, include=FALSE, eval=FALSE}
mint_data <- list(protein = rbind(
                            assays(data_24h)$loess %>% t, 
                            assays(data_48h)$loess %>% t,
                            assays(data_72h)$loess %>% t))

mint_data$study <- c(rep("t1" ,dim(data_24h)[2]), 
                     rep("t2" ,dim(data_48h)[2]),
                     rep("t3" ,dim(data_72h)[2]))

mint_data$Y <-  c(data_24h$binarysecout180daycpcscore, 
                  data_48h$binarysecout180daycpcscore,
                  data_72h$binarysecout180daycpcscore )

mint_data$Y2 <-  c(data_24h$shockable, 
                  data_48h$shockable,
                  data_72h$shockable )
            
#saveRDS(mint_data, "../minData_cpc.rds")            
                        
#mint_cpc <- mint.plsda(X, Y2, keepX=list.keepX2, ncomp = 3)

```

## Select Number of principle componenets (PCs)

In the first step we tune the number of PCs that capture most of the variation on the outcome variable. 

```{r tune_pcs, fig.cap="Number of PCs capturing maximum variation in the data"}
mint_data <- readRDS("../minData_cpc.rds")
mint.plsda.res.perf = mint.plsda(X = mint_data$protein, Y = mint_data$Y, study = mint_data$study, ncomp = 5)

set.seed(2543)  # for reproducible result in this example
perf.mint.plsda.cell <- perf(mint.plsda.res.perf, validation = "Mfold", folds = 5, 
                  progressBar = FALSE, auc = TRUE) 

plot(perf.mint.plsda.cell, col = color.mixo(5:7))
```
It is evident that PC1 captures most of the variation in the data. Therefore we can use 2 PCs to capture variation in PC1. PC2 will be helpful in plotting.

## Individual plot based PC1 and PC2

```{r plot_individuals}
mint.plsda.res = mint.plsda(X = mint_data$protein, Y = mint_data$Y, study = mint_data$study, ncomp = 2)
#mint.plsda.res # lists the different functions
plotIndiv(mint.plsda.res, legend = TRUE, title = 'MINT PLS-DA', 
          subtitle = 'CVD study', ellipse = T)
```

## Feature selection

In the next step we try to tune the number of feature from each of the time points that capture 

```{r tune_features, eval=FALSE}
debug(tune)
tune.mint = mixOmics::tune(X = mint_data$protein, Y = as.factor(mint_data$Y), study = as.factor(mint_data$study), nrepeat = 10, folds = 10, ncomp = 2, test.keepX=seq(1,100,1), method = 'mint.splsda', dist = "max.dist", progressBar = FALSE)

```
```{r}
DT::datatable(tune.mint$choice.keepX %>% as.data.frame())
```

## SPLSDA

Next we perform sparse PLSDA based 2 PCs and 15 and 19 proteins selected for PC1 and PC2 based one the tuning for feature selection.

```{r}
mint.splsda.res = mint.splsda(X = mint_data$protein, Y = as.factor(mint_data$Y), study = as.factor(mint_data$study), ncomp = 2,  
                              keepX = tune.mint$choice.keepX)

#selectVar(mint.splsda.res, comp = 1)
```

## Individual plot based on selected features on global dataset.

```{r}
plotIndiv(mint.splsda.res, study = 'global', legend = TRUE, title = 'MINT sPLS-DA', 
          subtitle = 'Global', ellipse=T)
```

### Individual plot for each of the timepoints

```{r}
plotIndiv(mint.splsda.res, study = 'all.partial',  title = 'MINT sPLS-DA', 
          subtitle = c("24h", "48h", "72h"))
```
### Correlation circle plot

The correlation circle plot shows the relation ship between selected variables from each of the time points on separation at two dimensional plot.

```{r}
plotVar(mint.splsda.res)
```

```{r}

plotLoadings(mint.splsda.res)
```


## Select Number of principle componenets (PCs)

In the first step we tune the number of PCs that capture most of the variation on the outcome variable. 

```{r tune_pcs, fig.cap="Number of PCs capturing maximum variation in the data"}
data_all <- readRDS("../data_sumExpfilImp.rds")
#data_all <- readRDS("../normalized_filtered_se.rds")
data_all_f <- data_all[, (data_all$shockable=="0" | data_all$shockable==1) & !is.na(data_all$shockable)]
mint_data_s <- list(protein = rbind(assays(data_all_f[, data_all_f$time=="-24-"])$loess %>% t,
                           assays(data_all_f[, data_all_f$time=="-48-"])$loess %>% t, 
                           assays(data_all_f[, data_all_f$time=="-72-"])$loess %>% t
                           ))

mint_data_s$study <- c(rep("t1" ,dim(data_all_f[, data_all_f$time=="-24-"])[2]), 
                     rep("t2" ,dim(data_all_f[, data_all_f$time=="-48-"])[2]),
                     rep("t3" ,dim(data_all_f[, data_all_f$time=="-72-"])[2]))

mint_data_s$Y <-  c(data_all_f[, data_all_f$time=="-24-"]$shockable, 
                  data_all_f[, data_all_f$time=="-48-"]$shockable,
                 data_all_f[, data_all_f$time=="-72-"]$shockable )


shock.mint.plsda.res.perf = mint.plsda(X = mint_data_s$protein, Y = mint_data_s$Y, study = mint_data_s$study, ncomp = 5)

set.seed(2543)  # for reproducible result in this example
shock.perf.mint.plsda.cell <- perf(shock.mint.plsda.res.perf, validation = "Mfold", folds = 5, 
                  progressBar = FALSE, auc = TRUE) 

plot(shock.perf.mint.plsda.cell, col = color.mixo(5:7))
```
## Individual plot based PC1 and PC2

```{r plot_individuals}
shock.mint.plsda.res = mint.plsda(X = mint_data_s$protein, Y = mint_data_s$Y, study = mint_data_s$study, ncomp = 2)
#mint.plsda.res # lists the different functions
plotIndiv(shock.mint.plsda.res, legend = TRUE, title = 'MINT PLS-DA', 
          subtitle = 'CVD study: Shockable vs. Nonshockable', ellipse = T)
```
## Feature selection

In the next step we try to tune the number of feature from each of the time points that capture 

```{r tune_features, eval=FALSE}
#debug(tune)
tune.mint_shock = mixOmics::tune(X = mint_data_s$protein, Y = as.factor(mint_data_s$Y), study = as.factor(mint_data_s$study), nrepeat = 10, folds = 10, ncomp = 2, test.keepX=seq(1,100,1), method = 'mint.splsda', dist = "max.dist", progressBar = FALSE)

```

```{r}
DT::datatable(tune.mint_shock$choice.keepX %>% as.data.frame())
```

## SPLSDA

Next we perform sparse PLSDA based 2 PCs and 15 and 19 proteins selected for PC1 and PC2 based one the tuning for feature selection.

```{r}
shock.mint.plsda.res = mint.splsda(X = mint_data_s$protein, Y = as.factor(mint_data_s$Y), study = as.factor(mint_data_s$study), ncomp = 2,  
                              keepX = tune.mint_shock$choice.keepX)

#selectVar(mint.splsda.res, comp = 1)
```

## Individual plot based on selected features on global dataset.

```{r}
plotIndiv(shock.mint.plsda.res, study = 'global', legend = TRUE, title = 'MINT sPLS-DA', 
          subtitle = 'Global', ellipse=T)
```
```{r}
plotIndiv(shock.mint.plsda.res, study = 'all.partial',  title = 'MINT sPLS-DA', 
          subtitle = c("24h", "48h", "72h"))
```


```{r}
plotLoadings(shock.mint.plsda.res, study = 'global')
```


```{r, echo=FALSE, include=FALSE, eval=FALSE}
selectVar(diablo_cpc, comp=2)
#variables selected on component 1 for each block
t1 <- selectVar(diablo_cpc, comp = 1, block = c(1,2,3))$'t1'$value %>% rownames_to_column()%>% as.tibble() %>% mutate(time= rep(1, dim(.)[1]))
t2 <- selectVar(diablo_cpc, comp = 1, block = c(1,2,3))$'t2'$value %>% rownames_to_column()%>% as.tibble() %>% mutate(time= rep(2, dim(.)[1]))
t3 <- selectVar(diablo_cpc, comp = 1, block = c(1,2,3))$'t3'$value %>% rownames_to_column()%>% as.tibble() %>% mutate(time= rep(3, dim(.)[1]))
sig2list <- c(t1$rowname,t2$rowname, t3$rowname ) %>% unique
```
